requirements.txt
streamlit
pandas
matplotlib
seaborn
pillow
scikit-learn
